Extract separate view for the site visit empty view
